backend/requirements.txt
aniso8601==7.0.0
asgiref==3.6.0
attrs==22.2.0
CacheControl==0.12.11
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.2.18
django-allauth==0.53.0
django-cors-headers==3.14.0
django-filter==22.1
django-graphql-auth
django-graphql-jwt
firebase-admin
flake8==6.0.0
gcloud==0.18.3
google-api-core==2.11.0
google-api-python-client==2.85.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.0
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
graphene==2.1.9
graphene-django==2.15.0
graphene-file-upload==1.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
httplib2==0.22.0
idna==3.4
iniconfig==2.0.0
mccabe==0.7.0
msgpack==1.0.5
oauth2client==4.1.3
oauthlib==3.2.2
packaging==23.0
pluggy==1.0.0
promise==2.3
proto-plus==1.22.2
protobuf==4.22.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT
pyparsing==3.0.9
Pyrebase4
pytest==7.2.1
python-jwt
python3-openid==3.2.0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rsa==4.9
singledispatch==4.0.0
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.15
whitenoise
wrapt==1.15.0
firebase-admin
pytest-cov==4.0.0
pytest-django==4.5.2